pyproject.toml
[project]
name = "google-ads-api-developer-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Bobby Andris",email = "bobby.andris@avondaledecor.com"}
]
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "google-ads (>=28.4.1,<29.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)"
]
